moved backup into dashboard/cluster/[clusterId]/backup
