+++ __tests__/s3backup.test.ts
@@ -1,5 +1,5 @@
 import { checkFileExistsInS3, deleteFromS3, uploadToS3 } from "@/services/aws-s3-service";
-import { backupFile } from "@/app/backup/backup.action";
+import { backupFile } from "@/app/dashboard/cluster/[clusterId]/backup/backup.action";
 
 describe("s3 upload database tests", () => {
   let fileCleanup: string | null;
